Minor Bug Fixes
Added ability to enable S4_TRACE debug calls via the Makefile

Fixed the writing of the real space vector field to text files in both
the python interface located in main_python.c and also in S4.cpp. There
was a bug in the Simulation_Destroy destructor function that attempted
to free a non-NULL pointer.