Minor Bug Fixes
Added ability to enable S4_TRACE debug calls via the Makefile

Fixed the writing of the real space vector field to text files in both
the python interface located in main_python.c and also in S4.cpp. There
was a bug in the Simulation_Destroy destructor function that attempted
to free a pointer to a Python object. According to the docs at this
link: https://docs.python.org/3/c-api/arg.html#api-functions when using
the Python C-API function to parse Python strings into a C pointer to a
character string using the 's' format option, we do not need to free the
memory ourselves. The buffer is managed by the corresponding Python
object and the buffer shares the lifetime of that object.