Got Improved Formulation Running!
Added a shit ton of comments in a bunch of places in the code

Also finally got Weismann's improved formulation running! This does not
mean it's actually working yet, but I have all the modifications and
basic code in place. Now I just need to make sure the matrices I'm using
and the way I'm multiplying them is all correct.